moving local-storage and remember-me to frontend, since it depends on running in a browser – in turn, the relative-time impelmentation no longer depends on the local storage wrapper
