+++ src/packages/frontend/account/init.ts
@@ -11,7 +11,7 @@ import { AccountStore } from "./store";
 import { AccountTable } from "./table";
 import { init_dark_mode } from "./dark-mode";
 import { reset_password_key } from "../client/password-reset";
-import { hasRememberMe } from "@cocalc/util/remember-me";
+import { hasRememberMe } from "@cocalc/frontend/misc/remember-me";
 import { appBasePath } from "@cocalc/frontend/customize/app-base-path";
 
 export function init(redux) {
