+++ src/packages/frontend/frame-editors/code-editor/actions.ts
@@ -60,7 +60,7 @@ import {
   set_local_storage,
   get_local_storage,
   delete_local_storage,
-} from "@cocalc/util/misc";
+} from "@cocalc/frontend/misc/local-storage";
 
 import {
   ext2syntax,
