+++ src/packages/frontend/misc/local-storage.ts
@@ -3,95 +3,51 @@
  *  License: AGPLv3 s.t. "Commons Clause" – see LICENSE.md for details
  */
 
-/*
- * Typed wrapper around LocalStorage
- */
+import { LocalStorageLRU } from "@cocalc/local-storage-lru";
+import { is_valid_uuid_string } from "@cocalc/util/misc";
 
-// tests at startup if localStorage exists and works. if not or disabled, uses memory as a fallback.
+const RECENTLY_KEY = "__recent";
+const RECENTLY_KEEP = 64; // tradeoff: keys need storage as well
 
-import { appBasePath } from "@cocalc/frontend/customize/app-base-path";
-
-const LS: { [k: string]: string | undefined } = (function () {
-  let it_works = false;
-  try {
-    const test_key = "cocalc_test";
-    window.localStorage[test_key] = "foo";
-    it_works = localStorage[test_key] == "foo";
-    delete localStorage[test_key];
-  } catch (e) {
-    console.warn(`Local Storage init issue ${e} -- using memory fallback`);
-  }
-  if (it_works) {
-    return window.localStorage;
-  } else {
-    return (window["cocalc_LS_memory"] = {});
-  }
-})();
-
-export class CustomKey {
-  constructor(private key: string) {}
-  getKey() {
-    return this.key;
-  }
-}
-
-type Keys = string[] | string | CustomKey;
-
-function make_key(keys: Keys): string {
-  if (keys instanceof CustomKey) {
-    return keys.getKey();
-  } else {
-    const key = typeof keys == "string" ? keys : keys.join(".");
-    return [appBasePath, key].join("::");
-  }
-}
-
-// returns the deleted value or undefined in case of a problem
-export function del<T>(keys: Keys): T | undefined {
-  const key = make_key(keys);
-  try {
-    const val = get<T>(keys);
-    delete LS[key];
-    return val;
-  } catch (e) {
-    console.warn(`localStorage delete("${key}"): ${e}`);
-  }
-}
-
-// set an entry, and return true if it was successful
-export function set<T>(keys: Keys, value: T): boolean {
-  const key = make_key(keys);
-  try {
-    LS[key] = JSON.stringify(value);
-    return true;
-  } catch (e) {
-    console.warn(`localStorage set("${key}"): ${e}`);
-    return false;
-  }
-}
-
-export function get<T>(keys: Keys): T | undefined {
-  const key = make_key(keys);
-  try {
-    const val = LS[key];
-    if (val != null) {
-      return JSON.parse(val);
-    } else {
-      return undefined;
+function candidate(key: string): boolean {
+  // only remove keys which start with an UUID, or are specific to editor- or jupyter-editor-
+  const uuidlen = 36;
+  for (const prefix of ["", "editor-", "jupyter-editor-"]) {
+    const i = prefix.length;
+    const j = i + uuidlen;
+    if (key.startsWith(prefix) && is_valid_uuid_string(key.slice(i, j))) {
+      return true;
     }
-  } catch (e) {
-    console.warn(`localStorage get("${key}"): ${e}`);
-    del<T>(key);
   }
+  return false;
 }
 
-export function exists(keys: Keys): boolean {
-  const key = make_key(keys);
-  // distinction between browser's localStorage and the fallback object
-  if (LS === window.localStorage) {
-    // test against null, see spec: https://www.w3.org/TR/webstorage/#dom-storage-getitem
-    return window.localStorage.getItem(key) !== null;
-  } else {
-    return LS[key] !== undefined;
-  }
+// Wrapper around localStorage, so we can safely touch it without raising an
+// exception if it is banned (like in some browser modes) or doesn't exist.
+// See https://github.com/sagemathinc/cocalc/issues/237
+export const LS = new LocalStorageLRU({
+  recentKey: RECENTLY_KEY,
+  maxSize: RECENTLY_KEEP,
+  isCandidate: candidate,
+  fallback: true,
+});
+
+export function set_local_storage(key: string, val: string): void {
+  LS.set(key, val);
+}
+
+export function get_local_storage(key: string): string | null {
+  return LS.get(key);
+}
+
+export function delete_local_storage(key: string): void {
+  LS.delete(key);
+}
+
+export function has_local_storage(): boolean {
+  return LS.localStorageWorks();
+}
+
+export function local_storage_length(): number {
+  return LS.size();
 }
