+++ src/packages/frontend/project/warnings/oom.tsx
@@ -7,7 +7,7 @@ import { Alert, Button } from "../../antd-bootstrap";
 import { useEffect, useRedux, useState, useTypedRedux } from "../../app-framework";
 import { A, Icon } from "../../components";
 import { ALERT_STYLE } from "./common";
-import * as LS from "../../misc/local-storage";
+import * as LS from "../../misc/local-storage-typed";
 
 const OOM_ALERT_STYLE: React.CSSProperties = {
   ...ALERT_STYLE,
