+++ src/packages/util/misc.ts
@@ -63,13 +63,6 @@ export {
   YEAR,
 } from "./relative-time";
 
-export {
-  set_local_storage,
-  get_local_storage,
-  has_local_storage,
-  delete_local_storage,
-  local_storage_length,
-} from "./local-storage";
 
 import sha1 from "sha1";
 export { sha1 };
