+++ src/packages/util/package.json
@@ -30,7 +30,6 @@
     "."
   ],
   "dependencies": {
-    "@cocalc/local-storage-lru": "^1.0.8",
     "@types/uuid": "^8.3.1",
     "async": "^1.5.2",
     "async-await-utils": "^3.0.1",
